APNS notifications working via Firebase
On important caveat is that apns does not handle pure 'data messages'
from firebase as one would hope. It is advisable to use seperate
topics for iOS and android and send data messages to android and
notification messages to iOS. This was you get the same behaviour on
both platforms.

- We remove the static constructors from Common.uno as these wont be
  fired until the class is touched and that could be very late in
  startup (we need to be doing setup during iOS start)
- We add a static constructor to JS.uno as the class is touched during
  startup and we use this to call the Init method in Common.uno
- We disable firebase's method swizzling so we can handle it ourselves
- FireNotificationCallbacks now implements
  UNUserNotificationCenterDelegate so we can set up the system
  correctly for iOS 10 and higher
- RegisterForPushNotifications now registers for notifications in the
  way expected by firebase's documentation